back/requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2025.11.12
cffi==2.0.0
Django==4.1.7
django-cors-headers==3.14.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
outcome==1.2.0
Pillow==12.0.0
pycparser==2.21
PyJWT==2.6.0
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2022.7.1
selenium==4.39.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.3
trio==0.32.0
trio-websocket==0.12.2
tzdata==2022.7
urllib3==2.6.2
wsproto==1.2.0
gunicorn==20.1.0
psycopg2-binary==2.9.11